calculate cost of current move and next move at same time fix memory leaks make pc moves all cost 10 for this one make swimmer M instead of S added more delete functions to heap implement priority queue for turn orders add wait until next turn code

+++ data/heap.c
@@ -78,7 +78,6 @@ node *deleteMin(node *root)
         child = child->sibling;
     }
     node *newRoot = mergePairs(minNode->child);
-    free(minNode->data);
     free(minNode);
     return newRoot;
 }
@@ -94,7 +93,20 @@ void traverse(node *root)
     traverse(root->sibling);
 }
 
+void deleteAllData(node *root)
+{
+    if (root == NULL)
+        return;
+    free(root->data);
+    while (NULL != (root = deleteMin(root)))
+    {
+        free(root->data);
+    }
+}
+
 void deleteAll(node *root)
 {
+    if (root == NULL)
+        return;
     while (NULL != (root = deleteMin(root)));
 }
